Switched from Cantor Pair to Bit Hash

+++ MS_Project/CantorPair.h
@@ -8,11 +8,11 @@ class Coord;
 class CantorPair
 {
 public:
-	static int get_int(Position* pos);
-	static int get_int(AStarNode* node);
-	static int get_int(Position* pos_1, Position* pos_2);
-	static int get_int(Coord* coord, int depth);
-	static int get_int(Coord* coord);
+	static unsigned int get_int(Position* pos);
+	static unsigned int get_int(AStarNode* node);
+	static unsigned int get_int(Position* pos_1, Position* pos_2);
+	static unsigned int get_int(Coord* coord, unsigned short depth);
+	static unsigned int get_int(Coord* coord);
 };
 
 #endif
